feat: refactor profile page and add ProfileTabs component for improved employee information display
