+++ src/app/dashboard/profile/page.tsx
@@ -1,48 +1,25 @@
 'use client';
 
-import { getEmployee } from '@/api/employee/employee';
-import ProtectPage from '@/components/auth/ProtectPage';
+import { Tabs, TabsList, TabsTrigger, TabsContent } from '@/components/ui/tabs';
+import ProfileTabs from './ProfileTabs';
 import { HeaderTitle } from '@/components/HeaderTitle';
-import ProfileMenuItem from '@/components/ProfileMenuItem';
-import { Button } from '@/components/ui/button';
 import { Card, CardDescription, CardTitle } from '@/components/ui/card';
-import { Input } from '@/components/ui/input';
-import { Label } from '@/components/ui/label';
-import { Separator } from '@/components/ui/separator';
-import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs';
-import { useApp } from '@/contexts/AppContext';
-import { Gender, GENDER_LABELS } from '@/enums/genderEnum';
-import { Marital, MARITAL_LABELS } from '@/enums/maritalEnum';
-import { TypeWork, TYPEWORK_LABELS } from '@/enums/typeWorkEnum';
-import { UserRole } from '@/enums/userRolesEnum';
-import type { EmployeeDto } from '@/models/dto/employeeDto';
-import { Briefcase, Mail, Paperclip, Pencil } from 'lucide-react';
 import Image from 'next/image';
-import { useRouter } from 'next/navigation';
+import { getEmployee } from '@/api/employee/employee';
 import { useEffect, useState } from 'react';
-import { TextBorder } from '@/components/BoderText';
-
-interface TabsInformation {
-  value: string;
-  label: string;
-}
-
-const tabsInformation: TabsInformation[] = [
-  { value: 'personal-info', label: 'Thông tin cá nhân' },
-  { value: 'professional-information', label: 'Thông tin nâng cao' },
-  { value: 'documents', label: 'Tài liệu' },
-  { value: 'account-access', label: 'Tài khoản truy cập' },
-];
-
-const documentsList = ['Tải hồ sơ CV', 'Tải hồ sơ học vấn', 'Tải hồ sơ chứng chỉ tiếng anh', 'Tải hồ sơ chứng chỉ khác'];
-
+import { EmployeeDto } from '@/models/dto/employeeDto';
+import { useRouter } from 'next/navigation';
+import { useApp } from '@/contexts/AppContext';
+import { Briefcase, Mail, Pencil } from 'lucide-react';
+import { Button } from '@/components/ui/button';
+import ProfileMenuItem from '@/components/ProfileMenuItem';
+import { UserRole } from '@/enums/userRolesEnum';
+import ProtectPage from '@/components/auth/ProtectPage';
 const ProfilePage = () => {
   const { userAccount } = useApp();
   const router = useRouter();
 
   const [employee, setEmployee] = useState<EmployeeDto | null>(null);
-  const [tab, setTab] = useState<string>(tabsInformation[0].value);
-  const [documents, setDocuments] = useState<string[]>(Array(documentsList.length).fill(''));
   useEffect(() => {
     const fetchEmployeeDetail = async () => {
       if (userAccount) {
@@ -50,35 +27,20 @@ const ProfilePage = () => {
         console.log('res', res);
         console.log('resss', userAccount?.employeeId);
         setEmployee(res);
-        setDocuments(res.document);
       }
     };
 
     fetchEmployeeDetail();
   }, [userAccount]);
 
-  const getGenderLabel = (gender?: string) => {
-    if (!gender) return '';
-    return GENDER_LABELS[gender as Gender] ?? '';
-  };
-
-  const getMartialLabel = (martial?: string) => {
-    if (!martial) return '';
-    return MARITAL_LABELS[martial as Marital] ?? '';
-  };
-
-  const getTypeLabel = (type?: string) => {
-    if (!type) return '';
-    return TYPEWORK_LABELS[type as TypeWork] ?? '';
-  };
-
   return (
     <>
       <HeaderTitle text="Thông tin nhân viên" subText="Thông tin chi tiết của nhân viên" />
-      <Card className="">
+      <Card>
         <div className="px-6 py-0 ">
-          <div className="flex md:flex-row md:items-end md:justify-between">
-            <div className="flex items-start space-x-4 ">
+          <div className="flex flex-col gap-3 md:flex-row md:items-center md:justify-between">
+            {/* Left Content */}
+            <div className="flex md:flex-row md:items-start space-x-4">
               {/* Profile Image */}
               <Image
                 src={(employee?.avatarUrl as string) ?? '/assets/default-avatar.png'}
@@ -109,7 +71,7 @@ const ProfilePage = () => {
 
             {/* Edit Button */}
             <Button
-              className="w-max"
+              className="w-full md:w-auto lg:w-auto"
               onClick={() => {
                 router.push(`/dashboard/employees/info-update/${userAccount?.employeeId}`);
               }}
@@ -118,135 +80,19 @@ const ProfilePage = () => {
               Edit Profile
             </Button>
           </div>
-
-          <Separator className="my-6" />
-
-          <div className="">
-            <div className="flex justify-between items-start gap-3">
-              <div className="w-auto">
+          
+            {/* Profile Menu */}
+           
+              <div className="mt-2">
                 <ProfileMenuItem />
               </div>
+            
+            {/* Tabs */}
 
-              <div className="w-full">
-                <Tabs
-                  defaultValue="personal-info"
-                  className="w-full mr-auto"
-                  value={tab}
-                  onValueChange={(value) => setTab(value)}
-                >
-                  <TabsList className="mb-3 w-[350px] md:w-auto flex justify-start overflow-x-auto ">
-                    {tabsInformation.map((tab) => (
-                      <TabsTrigger
-                        key={tab.value}
-                        value={tab.value}
-                        className={`
-            relative px-6 py-2 text-sm font-medium text-muted-foreground
-            rounded-none bg-transparent
-            transition-all duration-200 ease-in-out
-            hover:text-foreground hover:bg-accent/30
-            data-[state=active]:text-primary
-            data-[state=active]:bg-transparent
-            after:absolute after:bottom-0 after:left-0 after:h-[2px] after:w-full after:scale-x-0 after:bg-primary after:transition-transform after:duration-200
-            data-[state=active]:after:scale-x-100
-          `}
-                      >
-                        {tab.label}
-                      </TabsTrigger>
-                    ))}
-                  </TabsList>
-                  {/* Personal-info */}
-                  <TabsContent value="personal-info">
-                    <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
-                      <TextBorder label="Họ" value={employee?.lastname ?? ''} />
-                      <TextBorder label="Tên" value={employee?.firstname ?? ''} />
-                      <TextBorder label="Căn Cước Công Dân" value={employee?.identityCard ?? ''} />
-                      <TextBorder label="Email" value={employee?.email ?? ''} />
-                      <TextBorder label="Điện thoại" value={employee?.phone ?? ''} />
-                      <TextBorder
-                        label="Ngày sinh"
-                        value={employee?.birthday ? new Date(employee.birthday).toLocaleDateString('vi-VN') : ''}
-                      />
-                      <TextBorder label="Giới tính" value={getGenderLabel(employee?.gender)} />
-                      <TextBorder label="Quốc gia" value={employee?.country ?? ''} />
-                      <TextBorder label="Tình trạng hôn nhân" value={getMartialLabel(employee?.marital)} />
-                      <TextBorder label="Địa chỉ thường trú" value={employee?.address ?? ''} />
-                    </div>
-                  </TabsContent>
-                  {/* Professional-info */}
-                  <TabsContent value="professional-information">
-                    <div className="grid grid-cols-1 md:grid-cols-2 gap-4 mb-4">
-                      <TextBorder label="Mã nhân viên" value={employee?.employeeCode ?? ''} />
-                      <TextBorder label="Username" value={employee?.username ?? ''} />
-                      <TextBorder label="Chức vụ" value={getTypeLabel(employee?.type)} />
-                      <TextBorder label="Vị trí hiện tại" value={employee?.designation ?? ''} />
-                      <TextBorder label="Phòng ban" value={employee?.departmentId?.name ?? ''} />
-                      <TextBorder
-                        label="Ngày tham gia công ty"
-                        value={employee?.joinDate ? new Date(employee.joinDate).toLocaleDateString('vi-VN') : '--/--'}
-                      />
-                      <TextBorder
-                        label="Ngày bắt đầu làm việc"
-                        value={employee?.workingDate ? new Date(employee.workingDate).toLocaleDateString('vi-VN') : '--/--'}
-                      />
-                    </div>
-                  </TabsContent>
-                  {/* Documents */}
-                  <TabsContent value="documents">
-                    <div className="grid grid-cols-1 md:grid-cols-2 gap-5 mb-5">
-                      {documentsList.map((label, index) => (
-                        <div key={index} className="flex flex-col gap-2">
-                          <Label>{label}</Label>
-                          {documents[index] ? (
-                            (() => {
-                              const strUrl = documents[index];
-                              const fileUrl = strUrl.split('*')[1];
-                              const fileName = strUrl.split('*')[0];
-
-                              return (
-                                <div className="flex items-center justify-between bg-muted px-3 py-2 rounded mt-2">
-                                  <div className="flex gap-2">
-                                    <Paperclip size={20} className="text-primary" />
-                                    <a
-                                      href={fileUrl}
-                                      target="_blank"
-                                      rel="noopener noreferrer"
-                                      className="text-sm text-blue-600 dark:text-blue-400 hover:underline truncate"
-                                    >
-                                      {fileName}
-                                    </a>
-                                  </div>
-                                </div>
-                              );
-                            })()
-                          ) : (
-                            <div className="flex items-center justify-between bg-muted px-3 py-2 rounded mt-2">
-                              <div className="flex gap-2">
-                                <Paperclip size={20} className="text-primary" />
-                                <span className="text-sm text-gray-500 dark:text-gray-400">Chưa có tài liệu</span>
-                              </div>
-                            </div>
-                          )}
-                        </div>
-                      ))}
-                    </div>
-                  </TabsContent>
-                  {/* Account-access */}
-                  <TabsContent value="account-access">
-                    <div className="grid grid-cols-1 md:grid-cols-2 gap-4 mb-4">
-                      <TextBorder label="Email / Tài khoản đăng nhập" value={employee?.email ?? ''} />
-                      <TextBorder label="ID tài khoản Github" value={employee?.githubId ?? ''} />
-                      <TextBorder label="ID tài khoản Microsoft Teams" value={employee?.microsoftTeamId ?? ''} />
-                      <TextBorder label="ID tài khoản Slack" value={employee?.slackId ?? ''} />
-                    </div>
-                  </TabsContent>
-                </Tabs>
-              </div>
-            </div>
-          </div>
+         
         </div>
       </Card>
     </>
   );
-};
-
+}
 export default ProtectPage(ProfilePage, { allowedRoles: [UserRole.ADMIN, UserRole.HR, UserRole.EMPLOYEE] });
