+++ src/constants/nav-link/nav-link-items.ts
@@ -15,7 +15,6 @@ import {
   Users2,
   Workflow,
 } from 'lucide-react';
-
 const humanManageitems = [
   {
     title: 'Dashboard',
